Configure lldb to support /proc/self/cwd paths (#4992)
On my machine, the LLVM debug information seems to use source file paths
relative to `/proc/self/cwd`, which is the current working directory of
whatever process refers to it. Consequently, VSCode can't find those
files, because it has a different current working directory. This change
enables VSCode to rewrite those paths to a usable form.